Move tabs to components
