+++ apps/web/components/apps/layouts/InstalledAppsLayout.tsx
@@ -4,8 +4,7 @@ import AppCategoryNavigation from "@calcom/app-store/_components/AppCategoryNavi
 import { InstalledAppVariants } from "@calcom/app-store/utils";
 import { trpc } from "@calcom/trpc/react";
 import { Icon, Shell } from "@calcom/ui";
-import type { HorizontalTabItemProps } from "@calcom/ui/v2/core/navigation/tabs/HorizontalTabItem";
-import type { VerticalTabItemProps } from "@calcom/ui/v2/core/navigation/tabs/VerticalTabItem";
+import type { HorizontalTabItemProps, VerticalTabItemProps } from "@calcom/ui/";
 
 const tabs: (VerticalTabItemProps | HorizontalTabItemProps)[] = [
   {
