+++ packages/ui/components/index.ts
@@ -56,3 +56,5 @@ export { List, ListItem, ListItemText, ListItemTitle, ListLinkItem } from "./lis
 export type { ListItemProps, ListProps } from "./list";
 export { HeadSeo } from "./head-seo";
 export { Skeleton, SkeletonAvatar, SkeletonButton, SkeletonContainer, SkeletonText } from "./skeleton";
+export { HorizontalTabs, HorizontalTabItem, VerticalTabs, VerticalTabItem } from "./navigation";
+export type { HorizontalTabItemProps, NavTabProps, VerticalTabItemProps } from "./navigation";
