+++ packages/ui/components/navigation/index.tsx
@@ -1,0 +1,7 @@
+export { default as HorizontalTabItem } from "./tabs/HorizontalTabItem";
+export type { HorizontalTabItemProps } from "./tabs/HorizontalTabItem";
+export { default as HorizontalTabs } from "./tabs/HorizontalTabs";
+export type { NavTabProps } from "./tabs/HorizontalTabs";
+export { default as VerticalTabItem } from "./tabs/VerticalTabItem";
+export type { VerticalTabItemProps } from "./tabs/VerticalTabItem";
+export { default as VerticalTabs } from "./tabs/VerticalTabs";
