+++ packages/ui/components/navigation/tabs/HorizontalTabs.tsx
@@ -1,7 +1,5 @@
 import HorizontalTabItem, { HorizontalTabItemProps } from "./HorizontalTabItem";
 
-export { HorizontalTabItem };
-
 export interface NavTabProps {
   tabs: HorizontalTabItemProps[];
   linkProps?: HorizontalTabItemProps["linkProps"];
