+++ packages/ui/v2/core/index.ts
@@ -23,10 +23,6 @@ export { Checkbox, DatePicker, FormStep } from "./form";
 export { default as LinkIconButton } from "./LinkIconButton";
 export { default as Loader } from "./Loader";
 export { default as MeetingTimeInTimezones } from "./MeetingTimeInTimezones";
-export type { HorizontalTabItemProps } from "./navigation/tabs/HorizontalTabItem";
-export { default as HorizontalTabs, HorizontalTabItem } from "./navigation/tabs/HorizontalTabs";
-export type { VerticalTabItemProps } from "./navigation/tabs/VerticalTabItem";
-export { default as VerticalTabs, VerticalTabItem } from "./navigation/tabs/VerticalTabs";
 export { default as showToast } from "./notifications";
 export { default as SettingsToggle } from "./SettingsToggle";
 export { default as Shell } from "./Shell";
